refactor: merge feats_grants into feats module
Move grant application into core/feats.py and extract shared proficiency
token parsing to core/grant_mechanics.py for proficiencies and feats.
